+++ src/blocks/index.ts
@@ -1,4 +1,5 @@
 export * from "./cta-basic";
 export * from "./hero-basic";
 export * from "./image-text-card";
+export * from "./layout-elements";
 export * from "./text-image";